requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiohttp==3.9.1
alembic==1.12.1
pydantic==2.5.2
pytest==7.4.3
black==23.11.0
flake8==6.1.0
python-multipart==0.0.6
jinja2==3.1.2
aiosqlite==0.19.0
httpx==0.25.2 